infrastructure/webhook/requirements.txt
chalice >= 1.9.1
boto3 >= 1.11.9
gitpython >= 3.1.7